requirements.txt
streamlit>=1.39.0
openai>=1.55.0
pydantic>=2.9.0
python-dotenv>=1.0.1
vaderSentiment>=3.3.2
langdetect>=1.0.9
